sim particles created out of simulator
